SearchList now only shows games that have images
